requirements.txt
streamlit>=1.32.0
snowflake-snowpark-python>=1.5.0
snowflake-connector-python>=3.5.0
pandas>=2.2.0
plotly>=5.18.0
pypdf>=4.0.0
pdfplumber>=0.10.3
python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.4.0 